cleanup diff a bit
