+++ preview/assets/main.css
@@ -6,7 +6,7 @@
   --highlight-color-main: #dd7230;
   --highlight-color-secondary: #eb5160;
   --highlight-color-tertiary: #2b7fff;
-  --dx-content-width: 1120px;
+  --dx-content-width: 1360px;
 }
 
 html,
